software/requirements.txt
astropy==6.1.7
astropy-iers-data==0.2026.4.6.0.54.57
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
fonttools==4.62.1
future==1.0.0
h5py==3.16.0
iso8601==2.1.0
kiwisolver==1.5.0
matplotlib==3.10.8
numpy==2.2.6
packaging==26.0
pillow==12.2.0
pyerfa==2.0.1.5
pyparsing==3.3.2
PyQt5==5.15.11
PyQt5-Qt5==5.15.18
PyQt5_sip==12.18.0
pyqtgraph==0.14.0
pyserial==3.5
python-dateutil==2.9.0.post0
PyYAML==6.0.3
scapy==2.7.0
six==1.17.0
